Golang SDK (#218)
## Type of change

<!-- (mark with an `X`) -->

```
- [ ] Bug fix
- [ x] New feature development
- [ ] Tech debt (refactoring, code cleanup, dependency upgrades, etc)
- [ ] Build/deploy pipeline (DevOps)
- [ ] Other
```

## Objective

The objective is to develop a Golang wrapper for the Bitwarden client
that interacts with a native rust library. This is to ensure that a
consistent interface is provided, and the details of the Bitwarden rust
library can be abstracted away.

## Code changes
BitwardenClient.go

Defines the BitwardenClient struct with dependencies and submodules for
handling projects and secrets.
Implements the NewBitwardenClient function that creates a new client.
    Implements methods for access token login and cleanup.

BitwardenLibrary_native.go

Implements the BitwardenLibrary interface using cgo for interacting with
the native C library.
Provides methods for initialization, memory release, and running
commands on the client.

BitwardenLibrary_custom.go

An alternative implementation of the BitwardenLibrary interface,
possibly for testing or alternative backends. The code is currently the
same as BitwardenLibrary_native.go.

CommandRunner.go

Implements the CommandRunner struct that takes care of running commands
via the library interface.
    Centralizes command-running logic.

Projects.go

Implements a Projects struct with methods to create, list, get, update,
and delete projects.
    Utilizes the CommandRunner interface to run these commands.

Secrets.go

Implements a Secrets struct with methods to create, list, get, update,
and delete secrets.
    Utilizes the CommandRunner interface to run these commands.


*Notes*

- The code is designed to be modular, separating concerns into different
components (like Projects, Secrets, and CommandRunner).
- Interfaces are used where possible to make it easier to mock these
components for testing.
- The native C API is encapsulated neatly, making it easier to switch
out or update the underlying library without affecting the rest of the
code.

---------

Co-authored-by: Boris Bujak <boris.bujak@carvana.com>
Co-authored-by: Boris Bujak <bujakboris@gmail.com>