Reduce the number of CI jobs that are run on PRs (#375)
## Type of change
```
- [ ] Bug fix
- [ ] New feature development
- [ ] Tech debt (refactoring, code cleanup, dependency upgrades, etc)
- [x] Build/deploy pipeline (DevOps)
- [ ] Other
```

## Objective
With the recent addition of all the SDK language bindings, we've added a
lot of CI jobs on PRs that are unnecessary and slow down builds. By
switching them to run only on merge we should save some CI time.

Ultimately, we'd want some way to make it so that the `Build Rust Cross
Platform` get only called once and reused for every language SDK, but
for now this should help.