Change attachment key to EncString (#449)
## Type of change
```
- [x] Bug fix
- [ ] New feature development
- [ ] Tech debt (refactoring, code cleanup, dependency upgrades, etc)
- [ ] Build/deploy pipeline (DevOps)
- [ ] Other
```

## Objective
Change `AttachmentView` key to be an `EncString`, and skip decrypting
it. This is to avoid a uniffi issue with the representation of Vec<u8>
on Android.

Note that this is very similar to what we're doing with Sends, where we
don't decrypt the key on the model either until we're ready to decrypt
the file.