Split EncString for asymmetric support (#448)
## Type of change

```
- [ ] Bug fix
- [ ] New feature development
- [x] Tech debt (refactoring, code cleanup, dependency upgrades, etc)
- [ ] Build/deploy pipeline (DevOps)
- [ ] Other
```

## Objective
Split `EncString` into an asymmetric variant. I've left the symmetric
variant with the name `EncString` as it's the most common variant to
avoid major changes.

This is somewhat related to #447 and should be merged first so that PR
can take advantage of the symm/asymm separation.

I've moved the `EncString` files into their own folder to organize the
crypto directory a bit better too.