Sm 989 create build pipeline for the python language wrapper (#369)
## Type of change

- [ ] Bug fix
- [ ] New feature development
- [x] Tech debt (refactoring, code cleanup, dependency upgrades, etc)
- [x] Build/deploy pipeline (DevOps)
- [ ] Other

## Objective

Package the Python SDK so that it can be published to
[PyPI.org](pypi.org).

## Code changes

- **`.github/workflows/build-python-wheels.yml`:** create wheels for
Linux, macOS, and Windows (AMD64-only). A sdist is also created for
other platforms.
- **`languages/python/BitwardenClient/__init__.py`:** export functions
and classes for autocompletion in IDEs and the interactive Python
interpreter
- **`languages/python/pyproject.toml`**: replaces `setup.py`, which is
no longer officially supported. This uses Maturin to build the Rust
shared-object library that will be bundled with the SDK.
- **`languages/python/setup.py`**: replaced by `pyproject.toml`

## Before you submit

- Please add **unit tests** where it makes sense to do so (encouraged
but not required)

---------

Co-authored-by: Daniel García <dani-garcia@users.noreply.github.com>