[PM-5692] Extract generators to separate crate (#511)
Continuation on #402. This PR extracts the generators to their own crate
to establish clear boundaries between the code. There is still some
logic in the client to expose generators that we may want to revisit.

Also extracts explicit errors for the different generators for better
error handling downstream when/if desired.