[PM-3438] Vault exports (#561)
Vault export implementation.

New crate `bitwarden-exporters` which defines it's own data structure,
isolating it from the other code areas. It primarily focuses on
converting a data model to csv and json. Currently all data has to be
decrypted independently of what is required for the export which is an
area of improvement for the future.