Update swagger generated API crates (#590)
## Type of change
```
- [ ] Bug fix
- [ ] New feature development
- [x] Tech debt (refactoring, code cleanup, dependency upgrades, etc)
- [ ] Build/deploy pipeline (DevOps)
- [ ] Other
```

## Objective
Update swagger generator to version 7.2.0, pull in the new templates and
generate the new API crates.

I've also updated the build instructions to use .NET 8, as is required
in the latest server code.

The version of the server code used to generate this is
https://github.com/bitwarden/server/commit/fc1d7c70592cfbc65e61c047bcc138850d91a55b
I've had to make some small modifications to the server code to get it
working, ~~and I'll link the PR here shortly.~~
https://github.com/bitwarden/server/pull/3760

To make this easier to review, I've split this into three commits:
- b42f189924a770aa127b88478876548e8beb3879: Update the swagger and
dotnet versions, and pull in the new templates as-is. Note that this
won't generate working code.
- 9727d462924f9e533c0e3608ff0ba8aaeb0a2124: Patched the provided
templates to make it work with our code. This is in a separate commit to
make it easier to see what we've changed and hopefully make it easier in
the future to update these templates. The changes are:
    - Ignore warnings by default 
    - Use Uuids instead of strings
    - Use serde_repr to support int enums
    - Disabled default features for reqwest (to remove default tls)
- Removing double optionals (This could be reverted once we're better at
marking everything as required in the server)
- ~~f9b63606477d0f0f987b8534ef6e6755c76003a2: New autogenerated swagger
code, no manually edited files here.~~ Moved to a separate PR (#593)