Scanning (#630)
## Objective

Introduces a new GitHub Actions workflow named `Scan` in the
`.github/workflows/scan.yml` file. The workflow is triggered on manual
dispatch, push to the `main`, `rc`, and `hotfix-rc` branches, and when a
pull request is opened or synchronized. The workflow includes two jobs:
`sast` and `quality`. The `sast` job runs a Static Application Security
Testing (SAST) scan using Checkmarx and uploads the results to GitHub.
The `quality` job runs a quality scan using SonarCloud.

*
[`.github/workflows/scan.yml`](diffhunk://#diff-246cd0c2f7db532638dd80a92ac011f49b3d26038983a4c0169ea8f8a5c39280R1-R71):
Added a new GitHub Actions workflow named `Scan`. This workflow is
triggered on manual dispatch, push to the `main`, `rc`, and `hotfix-rc`
branches, and when a pull request is opened or synchronized. The
workflow includes two jobs: `sast` and `quality`. The `sast` job runs a
Static Application Security Testing (SAST) scan using Checkmarx and
uploads the results to GitHub. The `quality` job runs a quality scan
using SonarCloud.

## Before you submit

- Please add **unit tests** where it makes sense to do so