Add sensitive types to SymmetricCryptoKey (#672)
## Type of change
```
- [ ] Bug fix
- [ ] New feature development
- [x] Tech debt (refactoring, code cleanup, dependency upgrades, etc)
- [ ] Build/deploy pipeline (DevOps)
- [ ] Other
```

## Objective
This is a small subset of #536, which only contains the minimum code to
protect the import/export functions on SymmetricCryptoKey. It also
enables the from_base64 test on SymmetricCryptoKey as that passes now.

After this PR is merged I'll be expanding the use of Sensitive to other
parts of the codebase while adding some extra memory testing checks to
validate that it works for them.