[PM-7068] - Use a distroless container image for `bws` (#681)
## Type of change

- [ ] Bug fix
- [ ] New feature development
- [ ] Tech debt (refactoring, code cleanup, dependency upgrades, etc)
- [x] Build/deploy pipeline (DevOps)
- [x] Other

## Objective

Addresses [PM-7068](https://bitwarden.atlassian.net/browse/PM-7068).
Build the `bws` Docker image from an empty file system. This results in
a much smaller Docker image (~16MB, uncompressed) with a smaller threat
surface than bundling it with a distro.

## Code changes

- **./crates/bws/Dockerfile:** Use `scratch` for the final build stage.
This results in a distroless image that only contains our binary, the
libraries that it depends on, and the CA certificates needed for SSL to
work. The `ldd` line automatically determines what dependencies we need
to copy over so we don't have to manually maintain a list of them.

## Before you submit

- Please add **unit tests** where it makes sense to do so


[PM-7068]:
https://bitwarden.atlassian.net/browse/PM-7068?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ