Document developer tools (#723)
## Type of change
```
- [ ] Bug fix
- [ ] New feature development
- [ ] Tech debt (refactoring, code cleanup, dependency upgrades, etc)
- [ ] Build/deploy pipeline (DevOps)
- [x] Other
```

## Objective
Documented the tools we use for development. Sadly couldn't integrate
`nextest` into `bacon` as the output is too different and `bacon` relies
on parsing it to work. I still included it in the list for the cases
where devs run it manually.

I've also removed the `lto = true` instruction from bitwarden-napi, as
it was ignored in favor of the workspace setting and was producing a
warning.