Update uniffi to 0.27.x for async traits support (#756)
## Type of change
```
- [ ] Bug fix
- [ ] New feature development
- [x] Tech debt (refactoring, code cleanup, dependency upgrades, etc)
- [ ] Build/deploy pipeline (DevOps)
- [ ] Other
```

## Objective

Update uniffi crates to get async traits support for the passkeys API

## Code changes

This update doesn't have any breaking code changes but it requires
updating the Android compile SDK to 34, which also means updating the
Android Gradle Plugin and the Gradle Wrapper version. Personally, I also
had to update Android Studio to the latest version for it to work.
Presumably this won't matter to SDK users, as they will use the compiled
artifact directly.

While I was at it I also decided to update and cleanup the sdk-android
package dependencies.

I also updated the gradle wrapper version for the Java secrets-manager
bindings to keep them in sync, though this is not needed.