Fix ciphers & collections using wrong key when decrypting (#812)
In #738 we migrated away from `decrypt` and used `decrypt_with_key`
instead. However ciphers can be encrypted using either organization keys
or user key. The changes introduced meant we always used the users key
instead of conditionally based on the value of `organization_jd` in
`Cipher`.

The fix was to call `locate_key` on `Cipher` and `Collection`.

To properly write tests I needed to mock a user account and
organization. I took this moment to refactor our tests account and we
now have nice helpers for initializing the test@bitwarden.com account,
which also includes an organization.