Migrate fido logic to use explicit errors (#835)
This PR establishes explicit error types for the fido logic.

In order to extract the fido logic into a separate `bitwarden-fido`
crate, we need to remove the dependencies on the `bitwarden` logic. Most
of this is either the `Client` struct which is currently difficult, or
the `Error` and `Result` types which this PR focuses on.