Update PyO3 (#865)
Update PyO3 to version 0.21. Note that version 0.22 is already released
but `pyo3-log` is not compatible yet.

I've also noticed that we're not using `pyo3-asyncio` at all, as we only
expose a sync API on the Python libs, so I just removed it.

With this update and https://github.com/bitwarden/sdk/pull/864 we get
rid of any dependencies using `syn 0.x`, which should help a little bit
with build times.