Always deploy all rust packages (#845)
GitHub actions only supports 10 inputs to a workflow_dispatch. As we
keep adding packages we've exceeded this limit. Since we no longer need
the functionality to release individual packages this PR removes the
option all together and instead will always release all packages.

In case a package with the same version already exists in crates.io,
cargo release won't attempt to publish it and continue on with the other
packages.