[SM-1338] Add the ability to edit unassigned secrets (#906)
## 🎟️ Tracking

<!-- Paste the link to the Jira or GitHub issue or otherwise describe /
point to where this change is coming from. -->
https://bitwarden.atlassian.net/browse/SM-1338

## 📔 Objective

<!-- Describe what the purpose of this PR is, for example what bug
you're fixing or new feature you're adding. -->
With the introduction of individual secret permissions, it is possible
for a machine account to have **read, write** permission on an
unassigned secret.
Editing this unassigned secret should be supported by `bws`.
This removes the client side validation that restricted the edit of an
unassigned secret.