[SM-1096] Unify State Variable Naming (#895)
## 🎟️ Tracking

<!-- Paste the link to the Jira or GitHub issue or otherwise describe /
point to where this change is coming from. -->

https://bitwarden.atlassian.net/browse/SM-1096

## 📔 Objective

<!-- Describe what the purpose of this PR is, for example what bug
you're fixing or new feature you're adding. -->

The naming around state files and paths is inconsistent. This PR aims to
make the usage consistent.

Every mention of state for access token auth should follow this naming
logic:

`state_dir` <- This refers to the state directory, where state files are
stored. The sdk is currently only aware of state files, directory
handling is still being done in bws.
`state_file` <- This refers to a specific state file.

In our code we don't append naming to variables to indicate the type
(Hungarian Notation), so I thought it pertinent to truncate `_path` from
these, preferring more descriptive `state_dir` and `state_file` for the
names.

Thanks @dani-garcia for pointing this out
[here](https://github.com/bitwarden/sdk/pull/559#discussion_r1474654682).

⚠️ **Note** ⚠️ This also simplifies the `state_file_dir` key in the
`bws` config file to `state_dir`. There will be more changes to this
functionality with SM-1174, so I will wait to update the changelog for
the next bws version until all this is completed.

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes