Bundle platform-verifier (#1039)
We've run into some issues related to rustls-platform-verifier
distributing an android library through their crate, which works great
if you use platform-verifier within an application. However since we
distribute an sdk, the application have difficulties locating the
android specific platform-verifier library.

To resolve this we follow the method established in
https://github.com/rustls/rustls-platform-verifier/issues/115 and bundle
the classes with the library.